+++ tests/test_basics.py
@@ -1,3 +1,5 @@
+import pytest
+
 from async_btree import FAILURE, SUCCESS, ExceptionDecorator, node_metadata
 
 
@@ -22,10 +24,11 @@ def test_exception_decorator_falsy():
     assert repr(ExceptionDecorator(Exception("test"))) == repr(Exception("test"))
 
 
-def test_node_metadata_do_not_change_behavior(kernel):
+@pytest.mark.curio
+async def test_node_metadata_do_not_change_behavior():
     async def a_func():
         return 'a'
 
-    assert kernel.run(a_func) == 'a'
+    assert await a_func() == 'a'
     # no change on behavior
-    assert kernel.run(node_metadata()(a_func)) == 'a'
+    assert await node_metadata()(a_func)() == 'a'