+++ tests/test_map_filter.py
@@ -1,3 +1,5 @@
+import pytest
+
 from async_btree import afilter, amap
 
 
@@ -9,26 +11,29 @@ async def even(a):
     return a % 2 == 0
 
 
-def test_amap_on_iterable(kernel):
+@pytest.mark.curio
+async def test_amap_on_iterable():
     async def process():
         return [i async for i in amap(inc, [1, 2])]
 
-    assert kernel.run(process) == [2, 3]
+    assert await process() == [2, 3]
 
 
-def test_afilter_on_iterable(kernel):
+@pytest.mark.curio
+async def test_afilter_on_iterable():
     async def process():
         return [i async for i in afilter(even, [0, 1, 2, 3, 4])]
 
-    assert kernel.run(process) == [0, 2, 4]
+    assert await process() == [0, 2, 4]
 
 
-def test_afilter_amap_aiter(kernel):
+@pytest.mark.curio
+async def test_afilter_amap_aiter():
     async def process1():
         return [i async for i in afilter(even, amap(inc, [0, 1, 2, 3, 4]))]
 
     async def process2():
         return [i async for i in amap(inc, afilter(even, [0, 1, 2, 3, 4]))]
 
-    assert kernel.run(process1) == [2, 4]
-    assert kernel.run(process2) == [1, 3, 5]
+    assert await process1() == [2, 4]
+    assert await process2() == [1, 3, 5]